Re-organize botcode and decorators
