+++ app/botcode/investigation.py
