+++ app/ks_test/topic_analyzer.py
@@ -9,7 +9,7 @@ from scipy.stats import ks_2samp
 from pandas import DataFrame, read_csv, concat
 
 from app import DATA_DIR
-from app.datetime_decorators import to_ts, fmt_date
+from app.decorators.datetime_decorators import to_ts, fmt_date
 from app.bq_service import BigQueryService
 from app.ks_test.interpreter import interpret, PVAL_MAX
