+++ start/botcode-v2/ising_model_bot_detector.py
