+++ start/botcode-v2/networkClassifierHELPER.py
