+++ test/test_botcode.py
@@ -3,7 +3,7 @@ from networkx import DiGraph
 
 from app.botcode import (ALPHA, LAMBDA_1, LAMBDA_2, EPSILON,
     compute_link_energy, compile_energy_graph, parse_bidirectional_links)
-from app.workers.investigate_botcode import classify_bot_probabilities
+from app.botcode.investigation import classify_bot_probabilities
 from conftest import compile_mock_rt_graph
 
 def test_default_hyperparams():
